fix: prevent extra render when using `render` method
Example:
  css(node).set('y', 10).render()

The call to `set` will queue the update for the next frame.
But the call to `render` will handle the changes immediately.
Currently, the queued update still fires, which is fine, but
we should check `hasChanges` to avoid the superfluous call
to the `onRender` callback, which is pointless since the
`changedValues` array is always empty by that time.

To accomplish this, I'm hoisting the `render` method so it
can be passed to framesync, which means the `render` helper
function can be merged into the `render` method.